evaluator_service/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
spacy
python-dotenv
requests